+++ src/main/java/team/tripbeats/backend/repository/CurationRepository.java
@@ -3,5 +3,5 @@ package team.tripbeats.backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import team.tripbeats.backend.entity.Curation;
 
-public interface CurationRepository extends JpaRepository<Curation, Integer> {
+public interface CurationRepository extends JpaRepository<Curation, Long> {
 }
