+++ src/main/java/team/tripbeats/backend/repository/SpotRepository.java
@@ -3,5 +3,5 @@ package team.tripbeats.backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import team.tripbeats.backend.entity.Spot;
 
-public interface SpotRepository extends JpaRepository<Spot, Integer> {
+public interface SpotRepository extends JpaRepository<Spot, Long> {
 }
