+++ src/main/java/team/tripbeats/backend/repository/TripDayRepository.java
@@ -3,5 +3,5 @@ package team.tripbeats.backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import team.tripbeats.backend.entity.TripDay;
 
-public interface TripDayRepository extends JpaRepository<TripDay, Integer> {
+public interface TripDayRepository extends JpaRepository<TripDay, Long> {
 }
