+++ src/main/java/team/tripbeats/backend/repository/TripDaySpotRepository.java
@@ -3,5 +3,5 @@ package team.tripbeats.backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import team.tripbeats.backend.entity.TripDaySpot;
 
-public interface TripDaySpotRepository extends JpaRepository<TripDaySpot, Integer> {
+public interface TripDaySpotRepository extends JpaRepository<TripDaySpot, Long> {
 }
